% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffBindSampleTable.R
\name{createDiffBindSampleTable}
\alias{createDiffBindSampleTable}
\title{Title}
\usage{
createDiffBindSampleTable(
  tpDrHealthy,
  tpDrExp,
  sampleID,
  fileIDExp,
  fileIDHealthy,
  bamFiles = F,
  bamFilesHealthInput = F,
  bamFilesExpInput = F,
  tissue = NULL,
  replicate = NULL,
  bedExt = "_trim_bowtie2UPBlkLstRm_MACS2_peaks.narrowPeak",
  bamExt = "_trim_bowtie2UPBlkLstRm.sorted.bam",
  peakCaller = "narrow",
  bamDr = "alignments/",
  bedDr = "peakCalls/",
  cond1 = "Healthy",
  cond2 = "OC"
)
}
\arguments{
\item{bamFiles}{}
}
\value{

}
\description{
Title
}
