% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGeneFunctions.R
\name{findSignificantPeaks}
\alias{findSignificantPeaks}
\title{Title}
\usage{
findSignificantPeaks(peaks, FDR, pvalue)
}
\arguments{
\item{peaks}{- a list of the peaks}

\item{pvalue}{}
}
\value{

}
\description{
Title
}
