% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGeneFunctions.R
\name{generateGeneTable}
\alias{generateGeneTable}
\title{Title}
\usage{
generateGeneTable(peaks, genes, GeneInds, peakTable, fileName)
}
\arguments{
\item{fileName}{}
}
\value{

}
\description{
Title
}
