% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmScf.R
\name{rmScf}
\alias{rmScf}
\title{Remove scaffolds. This funciton removes all the scaffold chromosomes from a genomic ranges object and then removes the sequence factors}
\usage{
rmScf(GR)
}
\arguments{
\item{GR}{Genomic ranges object with scaffolds to remove}
}
\value{
Returns the same GR object with the scaffolds removed. Also sets the chromosome style to UCSC
}
\description{
Remove scaffolds. This funciton removes all the scaffold chromosomes from a genomic ranges object and then removes the sequence factors
}
